pyproject.toml
[tool.poetry]
name = "doughbot"
version = "0.1.0"
description = ""
authors = ["Brett Kolodny <brettkolodny@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "==1.4.2"
requests = "^2.24.0"
pyfp = "^0.1.3"
pymongo = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "<2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
